finished SQL scripts

+++ DumDum/DumDum/Models/JsonEntities/Troops/TroopsResponse.cs
@@ -12,8 +12,8 @@ namespace DumDum.Models.JsonEntities
         public string TroopType { get; set; }
         public int Level { get; set; }
         public int HP { get; set; }
-        public int Attack { get; set; }
-        public int Defence { get; set; }
+        public double Attack { get; set; }
+        public double Defence { get; set; }
         public int StartedAt { get; set; }
         public int FinishedAt { get; set; }
     }
